rag_poc_function/requirements.txt
google-cloud-aiplatform==1.64.0
google-cloud-secret-manager==2.21.0
google-cloud-storage==2.16.0
weaviate-client==4.5.4
requests==2.31.0
beautifulsoup4==4.12.3
tiktoken==0.7.0
python-dotenv==1.0.1
tenacity==8.2.3
